feat(consultation): add $350 consultation + monograph bundle feature
Complete implementation of professional consultation purchase flow:

Database:
- Add 'consultation' value to payment_tier enum (migration)
- Enables storing consultation purchases in assessments table

Features:
- /comprehensive-care page: consultation purchase landing page
- /consultation-success page: post-purchase confirmation with Dr. Carpentier contact
- Stripe checkout: handles $350 consultation + monograph bundle
- Webhook: records consultation tier, skips email (shows contact on page)
- Check-ins: consultation users see NO upsells (purchase-aware CTAs)
- Tier mapping: consultation tier grants monograph access

Bundle Details:
- $350 total price
- Includes: 45-min phone consultation + Comprehensive Monograph ($20 value)
- Contact: 254-393-2114, drcpainmd.com
- No email sent (all info on success page)
- Users still get check-ins for data collection

Feature Flag: FEATURE_BUNDLE_350 (currently disabled)

Testing: End-to-end flow tested in staging
Ready: Awaiting production feature flag enablement

+++ supabase/migrations/20250930_value_first_metrics_view.sql
@@ -239,8 +239,8 @@ SELECT
   ) as last_30d_overall_conversion,
 
   -- Time on results screen stats
-  ROUND(avg_time_seconds, 2) as avg_time_on_results_seconds,
-  ROUND(median_time_seconds, 2) as median_time_on_results_seconds,
+  ROUND(avg_time_seconds::numeric, 2) as avg_time_on_results_seconds,
+  ROUND(median_time_seconds::numeric, 2) as median_time_on_results_seconds,
   sessions_with_time_data
 
 FROM time_stats;
